Extract Places details fetch into shared helper
Moves the Google Places details fetch + transform out of the
api.places.details route into lib/places.server.ts so it can be reused
by the restaurant metadata refresh job.
